+++ pulsar/apps/tasks/backends/local.py
@@ -1,6 +1,4 @@
-from collections import deque
-
-from pulsar import send, command
+from pulsar import send, command, Queue, Empty, coroutine_return
 from pulsar.utils.pep import itervalues
 from pulsar.apps.tasks import backends, states
 from pulsar.apps.pubsub import PubSub
@@ -11,8 +9,8 @@ class TaskBackend(backends.TaskBackend):
     def put_task(self, task_id):
         return send(self.name, 'put_task', task_id)
     
-    def get_task(self, task_id=None, timeout=1):
-        return send(self.name, 'get_task', task_id, timeout)
+    def get_task(self, task_id=None):
+        return send(self.name, 'get_task', task_id, self.poll_timeout)
     
     def get_tasks(self, **filters):
         return send(self.name, 'get_tasks', **filters)
@@ -23,21 +21,31 @@ class TaskBackend(backends.TaskBackend):
     def delete_tasks(self, ids=None):
         return send(self.name, 'delete_tasks', ids)
     
-        
+
+#########################################################    INTERNALS
+
 class LocalTaskBackend(object):
     
     def __init__(self, name):
         self.pubsub = PubSub(name=name)
         self._tasks = {}
-        self.queue = deque()
+        self.queue = Queue()
     
     def put_task(self, task_id):
         if task_id in self._tasks:
             task = self._tasks[task_id]
             task.status = states.QUEUED 
-            self.queue.append(task.id)
-            return task.id
+            yield self.queue.put(task.id)
+            yield task.id
 
+    def get_task(self, task_id, timeout):        
+        if not task_id:
+            try:
+                task_id = yield self.queue.get(timeout)
+            except Empty:
+                coroutine_return()
+        yield self._tasks.get(task_id)
+        
     def save_task(self, task_id, **params):
         task = self._tasks.get(task_id)
         if task:
@@ -47,14 +55,6 @@ class LocalTaskBackend(object):
             task = backends.Task(task_id, **params)
             self._tasks[task.id] = task
         return task.id
-    
-    def get_task(self, task_id, timeout):        
-        if not task_id:
-            try:
-                task_id = self.queue.popleft()
-            except IndexError:
-                return None
-        return self._tasks.get(task_id)
 
     def delete_tasks(self, ids):
         ids = ids or list(self._tasks)
@@ -107,7 +107,7 @@ def put_task(request, task_id):
     return _get_tasks(request.actor).put_task(task_id)
 
 def _get_tasks(actor):
-    tasks = actor.params.TASKQUEUE_TASKS
+    tasks = getattr(actor, '_TASKQUEUE_TASKS', None)
     if tasks is None:
-        actor.params.TASKQUEUE_TASKS = tasks = LocalTaskBackend(name=actor.name)
+        actor._TASKQUEUE_TASKS = tasks = LocalTaskBackend(name=actor.name)
     return tasks