+++ tests/internet/eventloop.py
@@ -12,14 +12,14 @@ server_protocol = lambda: Connection(1, 0, EchoServerProtocol, None)
 
 class TestEventLoop(unittest.TestCase):
     
-    def __test_create_connection_error(self):
+    def test_create_connection_error(self):
         loop = get_event_loop()
         try:
             result = yield loop.create_connection(Protocol,'127.0.0.1', 9898)
         except socket.error:
             pass
         
-    def __test_start_serving(self):
+    def test_start_serving(self):
         protocol_factory = lambda : Connection()
         loop = get_event_loop()
         sockets = yield loop.start_serving(server_protocol,'127.0.0.1', 0)
@@ -35,7 +35,7 @@ class TestEventLoop(unittest.TestCase):
         self.assertFalse(handler)
         self.assertTrue(is_socket_closed(socket))
         
-    def __test_start_serving_ipv6(self):
+    def test_start_serving_ipv6(self):
         loop = get_event_loop()
         sockets = yield loop.start_serving(Protocol,'::1', 0)
         self.assertEqual(len(sockets), 1)